% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ECH_raw_data.R
\name{read_ECH_raw_data}
\alias{read_ECH_raw_data}
\title{read_ECH_raw_data}
\usage{
read_ECH_raw_data(my_csv)
}
\arguments{
\item{my_csv}{A csv file containing AVAYA ECH CMS R17 raw data exported from the ECH handler script}
}
\value{
The imported raw data as a tbl_df
}
\description{
read_ECH_raw_data
}
\examples{
my_data = read_ECH_raw_data("data-raw/R17_example")
View(my_data)
}

